Remove sso_saml2_toolkit module and refactor code to use sso_saml2
