+++ python/tank/authentication/sso_saml2/__init__.py
@@ -25,8 +25,6 @@ from .core.errors import (  # noqa
 # Classes
 from .sso_saml2 import SsoSaml2  # noqa
 
-from .sso_saml2_toolkit import SsoSaml2Toolkit  # noqa
-
 # Functions
 from .utils import (  # noqa
     get_saml_claims_expiration,
